more generic for subject
