+++ lib/oregon_digital/controlled_vocabularies/subject.rb
@@ -25,44 +25,15 @@ module OregonDigital
       end
 
       def fetch(*_args, &_block)
-        case self.class.query_to_vocabulary(rdf_subject.to_s)
+        vocabulary = self.class.query_to_vocabulary(rdf_subject.to_s)
+        case vocabulary
         when 'OregonDigital::ControlledVocabularies::Vocabularies::Itis'
-          fetch_and_store
+          store_statement(fetch_itis_statement(vocabulary, rdf_subject))
         when 'OregonDigital::ControlledVocabularies::Vocabularies::Ubio'
-          fetch_and_store
+          store_statement(fetch_ubio_statement(vocabulary, rdf_subject))
         else
           super
         end
-      end
-
-      private
-
-      def fetch_and_store
-        new_statement = statement
-        OregonDigital::Triplestore.triplestore_client.insert([new_statement])
-        self << statement
-      end
-
-      def statement
-        RDF::Statement.new(rdf_subject, RDF::Vocab::SKOS.prefLabel, subject_label)
-      end
-
-      def subject_label
-        case vocabulary.to_s
-        when 'OregonDigital::ControlledVocabularies::Vocabularies::Itis'
-          parse_json(vocabulary)
-        when 'OregonDigital::ControlledVocabularies::Vocabularies::Ubio'
-          parse_xml
-        end
-      end
-
-      def parse_json(vocabulary)
-        uri = vocabulary.as_query(rdf_subject.to_s)
-        JSON.parse(Faraday.get(uri) { |req| req.headers['Accept'] = 'application/json' }.body)['scientificName']['combinedName']
-      end
-
-      def parse_xml
-        Nokogiri::XML(Faraday.get(rdf_subject).body).at_xpath('/rdf:RDF/rdf:Description/dc:title/text()')
       end
     end
   end
